Move everything to com.gu.identity.frontend packages
