+++ app/com/gu/identity/frontend/controllers/Application.scala
@@ -1,12 +1,10 @@
-package controllers
+package com.gu.identity.frontend.controllers
 
 import javax.inject.Inject
 
-import play.api.i18n.Messages
-import play.api.i18n.Messages.Implicits._
-import play.api.Play.current
+import com.gu.identity.frontend.logging.Logging
+import com.gu.identity.frontend.views
 import play.api.mvc._
-import util.Logging
 
 class Application @Inject() extends Controller with Logging {
