+++ conf/application.conf
@@ -1,6 +1,6 @@
 include file("/etc/gu/identity-frontend.conf")
 
 play.http.session.secure=true
-play.http.filters = "filters.Filters"
+play.http.filters = "com.gu.identity.frontend.filters.Filters"
 
 play.i18n.langs = ["en"]
